assignment_2/airflow/requirements.txt
#redis
#openai[embeddings]
#sentence-transformers
google-cloud-storage==2.9.0
SQLAlchemy
python-dotenv
cloud-sql-python-connector
cloud-sql-python-connector[pg8000]
requests